Updated Woocommerce
They don't say anything about needing to install bourbon. They just include the import in the main woocommerce scss file. Bourbon is a ruby gem and installs its own folder in the woocommerce css folder. Then the path to the bourbon partial had to be changed and a force recompile was necessary.